FrontEnd, Api, Security & Arq Update
All modules are connected to an API wich handles all methods realted to the controller of the database and the serverside operations. This way we can migrate the api and fully customize and monitor its performance. The client mostly gets json data. Still working with SESSION array but mostly migrating to a mobile friendly enviroment.